refactor: update FilesUseCases tests and imports
- Refactored tests for FilesUseCases to use getObjectInformation instead of getMetadata.
- Updated mock implementations to reflect changes in the download blocking logic.
- Cleaned up imports in files.spec.ts and index.ts for better organization and clarity.
