+++ backend/src/useCases/objects/files/index.ts
@@ -14,22 +14,16 @@ import {
   UploadType,
   DownloadServiceOptions,
 } from '@auto-drive/models'
-import {
-  ObjectUseCases,
-  OwnershipUseCases,
-  SubscriptionsUseCases,
-} from '../index.js'
-import { uploadsRepository } from '../../repositories/uploads/uploads.js'
-import { BlockstoreUseCases } from '../uploads/blockstore.js'
-import { asyncIterableToPromiseOfArray } from '@autonomys/asynchronous'
-import { downloadService } from '../../services/download/index.js'
-import { FileGateway } from '../../services/dsn/fileGateway/index.js'
+import { ObjectUseCases, OwnershipUseCases } from '../index.js'
+import { uploadsRepository } from '../../../repositories/uploads/uploads.js'
+import { BlockstoreUseCases } from '../../uploads/blockstore.js'
 import { Readable } from 'stream'
 import { createLogger } from '../../../drivers/logger.js'
 import { ByteRange } from '@autonomys/file-caching'
 import { sliceReadable } from '../../../utils/readable.js'
 import { DBObjectFetcher, FileGatewayObjectFetcher } from './fetchers.js'
 import { composeNodesDataAsFileReadable } from './nodeComposer.js'
+import { SubscriptionsUseCases } from '../../users/subscriptions.js'
 
 const logger = createLogger('useCases:objects:files')
